backend/packages/hsm_core/pyproject.toml
[project]
name = "hsm-core"
version = "0.1.0"
description = "Shared domain and job utilities for HSM backend"
requires-python = ">=3.11"
dependencies = [
    "firebase-admin>=6.9.0",
    "google-auth>=2.0.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-storage>=2.18.0",
    "numpy>=1.26.0",
    "pandas[parquet]>=2.0.0",
    "pyarrow>=14.0.0",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.11.5",
    "rasterio>=1.4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hsm_core"]

[tool.ruff]
line-length = 88
target-version = "py311"
